modellar va viewsni hozircha checkout da tugatdik

+++ store/models.py
@@ -14,7 +14,7 @@ class Customer(models.Model):
 
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=False)
     image = models.ImageField(blank=True, null=True)
 
@@ -38,6 +38,7 @@ class Order(models.Model):
 
     def __str__(self):
         return str(self.id)
+
 
     @property
     def shipping(self):
